+++ packages/ui/core-components/src/lib/unsorted/viz/references/ReferencePoint/reference-point.d.ts
@@ -1,6 +1,6 @@
 // @ts-check
 
-import type { Writable } from 'svelte/store';
+import type { Writable, Readable } from 'svelte/store';
 import type { MarkPointComponentOption } from 'echarts';
 
 import type { ReferenceColor, Symbol } from '../types.js';
@@ -32,9 +32,10 @@ export type ReferencePointStoreValue = {
 	align?: 'left' | 'center' | 'right';
 	bold?: boolean;
 	italic?: boolean;
+	error?: string;
 };
 
-export type ReferencePointStore = Writable<ReferencePointStoreValue> & Readable<{ error?: string }>;
+export type ReferencePointStore = Writable<ReferencePointStoreState>;
 
 export type ReferencePointChartData = MarkPointComponentOption['data'][number] & {
 	evidenceSeriesType: 'reference_point';